+++ app/src/androidTest/java/com/cmput301f20t21/book_friends/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.example.book_friends;
+package com.cmput301f20t21.book_friends;
 
 import android.content.Context;
