+++ app/src/main/java/com/cmput301f20t21/book_friends/ui/home/HomeFragment.java
@@ -1,4 +1,4 @@
-package com.example.book_friends.ui.home;
+package com.cmput301f20t21.book_friends.ui.home;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -12,7 +12,7 @@ import androidx.fragment.app.Fragment;
 import androidx.lifecycle.Observer;
 import androidx.lifecycle.ViewModelProviders;
 
-import com.example.book_friends.R;
+import com.cmput301f20t21.book_friends.R;
 
 public class HomeFragment extends Fragment {
